RunPod uses CDN

+++ cloudprovider/runpod/init.sh
@@ -4,12 +4,10 @@ set -euo pipefail
 mkdir -p /etc/virtualpod
 echo "${GPU_PROVIDER_WIREPROXY_CONFIG}" | base64 -d > /etc/virtualpod/wireproxy.tpl
 
-curl -fsSL "${GPU_PROVIDER_WIREPROXY_URL}" -o /usr/bin/wireproxy
-curl -fsSL "${GPU_PROVIDER_AGENT_URL}" -o /container_agent
-curl -fsSL "${GPU_PROVIDER_WSTUNNEL_URL}" -o /usr/bin/wstunnel
-[ -n "${GPU_PROVIDER_PROMTAIL_URL:-}" ] && curl -fsSL "${GPU_PROVIDER_PROMTAIL_URL}" -o /usr/bin/promtail && chmod +x /usr/bin/promtail
-
-chmod +x /usr/bin/wireproxy /container_agent /usr/bin/wstunnel
+[ -e /usr/bin/wireproxy ] || { curl -fsSL "${GPU_PROVIDER_WIREPROXY_URL}" -o /usr/bin/wireproxy && chmod +x /usr/bin/wireproxy; }
+[ -e /usr/bin/container_agent ] || { curl -fsSL "${GPU_PROVIDER_AGENT_URL}" -o /usr/bin/container_agent && chmod +x /usr/bin/container_agent; }
+[ -e /usr/bin/wstunnel ] || { curl -fsSL "${GPU_PROVIDER_WSTUNNEL_URL}" -o /usr/bin/wstunnel && chmod +x /usr/bin/wstunnel; }
+[ -e /usr/bin/promtail ] || { [ -n "${GPU_PROVIDER_PROMTAIL_URL:-}" ] && curl -fsSL "${GPU_PROVIDER_PROMTAIL_URL}" -o /usr/bin/promtail && chmod +x /usr/bin/promtail; }
 
 unset AWS_WEB_IDENTITY_TOKEN_FILE
 export PIP_PROXY="http://127.0.0.1:3128"
